direction endpoints

+++ api/internal/cache/cache.go
@@ -3,15 +3,15 @@ package cache
 import "time"
 
 type RefreshToken interface {
-	Save(userID uint8, token string, ttl time.Duration) error
-	Get(userID uint8) (string, error)
-	Delete(userID uint8) error
+	Save(userID uint, token string, ttl time.Duration) error
+	Get(userID uint) (string, error)
+	Delete(userID uint) error
 }
 
 type Blacklist interface {
-	Save(userID uint8, accessToken string, ttl time.Duration) error
-	Get(userID uint8) error
-	Delete(userID uint8) error
+	Save(userID uint, accessToken string, ttl time.Duration) error
+	Get(userID uint) error
+	Delete(userID uint) error
 }
 
 type Cache struct {
