+++ api/internal/dto/id.go
@@ -1,5 +1,5 @@
 package dto
 
 type IDResponse struct {
-	ID uint8 `json:"id"`
+	ID uint `json:"id"`
 }
